move into a better package object
